feat: Refactor codebase and implement class specialization
This commit introduces a major refactoring of the codebase and implements the class evolution feature.

Refactoring:
- The monolithic `rpg.py` has been split into four distinct modules for better organization and maintainability:
  - `models.py`: Contains all data classes (Player, NPC, Item, etc.).
  - `world.py`: Handles loading the game world from data files.
  - `managers.py`: Contains game logic managers (SkillTreeManager, ClassManager, etc.) and state management functions (save/load).
  - `main.py`: The main entry point and game loop.
- All modules now use explicit, non-wildcard imports.
- `__all__` has been added to modules to define a clear public API.

Class Specialization Feature:
- Players are now prompted to choose a permanent class (Knight, Ranger, or Mage) upon reaching level 10.
- Each class grants unique base stat modifications and access to an exclusive pool of skills.
- The Player model now uses a derived stats system (`base_` stats + `recalculate_stats`) to deterministically apply modifiers from classes and skills.
- The save/load system has been updated to persist the player's `class_id` and base stats.
- Added migration logic to handle old save files, prompting players to choose a class if they are level 10+ and haven't specialized.
- Updated the test suite to reflect the new modular structure and verify all new functionality.
